pyproject.toml
[project]
name = "cyberbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex DeNuzzo",email = "adenuzzo@lclark.edu"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0.0"
dependencies = [
    "langchain-core>=0.1.27",
    "langchain-community>=0.0.24",
    "langchain-anthropic>=0.3.12",
    "langchain>=0.1.9",
    "langchain-huggingface>=0.1.2",
    "python-dotenv>=1.1.0",
    "faiss-cpu>=1.7.4",
    "unstructured (>=0.17.2,<0.18.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
